requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py38h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.5.18.1=ha878542_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.5.18.1=py38h578d9bd_0
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38h06a4308_0
cryptography=3.4.8=py38hd23ed53_0
cudatoolkit=11.3.1=h2bc3f7f_2
cupy=10.4.0=py38h405e1b6_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
fastrlock=0.8=pypi_0
ffmpeg=4.2.2=h20bf706_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py38h7f8727e_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
grpcio=1.42.0=py38hce63b2e_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py38h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9e=h7f8727e_0
kiwisolver=1.4.2=py38h295c915_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_0
libgomp=11.2.0=h1234567_0
libidn2=2.3.2=h7f8727e_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libstdcxx-ng=11.2.0=h1234567_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py38h06a4308_0
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py38hef1b27d_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=5.2.0=py38h7f8727e_2
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
numpy=1.22.3=py38he7a7128_0
numpy-base=1.22.3=py38hf524024_0
oauthlib=3.2.0=pyhd3eb1b0_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1o=h166bdaf_0
pillow=9.0.1=py38h22f2fdc_0
pip=22.1.1=pypi_0
protobuf=3.20.1=py38h295c915_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.1.0=py38h06a4308_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.11.0=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
setuptools=62.3.2=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.3=hc218d9a_0
tensorboard=2.9.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py38hca6d32c_0
tensorboard-plugin-wit=1.6.0=py_0
tk=8.6.11=h1ccaba5_1
torchaudio=0.11.0=py38_cu113
torchvision=0.12.0=py38_cu113
tornado=6.1=py38h27cfd23_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=py38h06a4308_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7f8727e_1
yarl=1.6.3=py38h27cfd23_0
zipp=3.8.0=py38h06a4308_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
